requirements.txt
python-dotenv>=1.0.0
mistralai>=1.0.0
PyPDF2>=3.0.0
nicegui>=3.0.0
